+++ playground/src/Playground.vue
@@ -6,6 +6,7 @@ import DirectiveSection from './components/sections/DirectiveSection.vue'
 import DynamicLoadersSection from './components/sections/DynamicLoadersSection.vue'
 import FallbackSection from './components/sections/FallbackSection.vue'
 import FormatSection from './components/sections/FormatSection.vue'
+import IntrospectionSection from './components/sections/IntrospectionSection.vue'
 import LazyLoadSection from './components/sections/LazyLoadSection.vue'
 import MemorySection from './components/sections/MemorySection.vue'
 import PartialBlocksSection from './components/sections/PartialBlocksSection.vue'
@@ -13,6 +14,7 @@ import PlaygroundFooter from './components/sections/PlaygroundFooter.vue'
 import PlaygroundHeader from './components/sections/PlaygroundHeader.vue'
 import PluralSection from './components/sections/PluralSection.vue'
 import ScopeSyncSection from './components/sections/ScopeSyncSection.vue'
+import SubtreeSection from './components/sections/SubtreeSection.vue'
 import WildcardSection from './components/sections/WildcardSection.vue'
 import { usePlaygroundHooksFeed } from './composables/usePlaygroundHooksFeed'
 
@@ -42,6 +44,8 @@ const { events: hookEvents, clear: clearHookEvents, probeTranslate } = usePlaygr
           <DynamicLoadersSection />
           <PluralSection />
           <FormatSection />
+          <IntrospectionSection />
+          <SubtreeSection />
         </div>
 
         <PlaygroundFooter />